+++ lib/post/widgets/comment_card.dart
@@ -135,7 +135,7 @@ class _CommentCardState extends State<CommentCard> {
                     return Padding(
                       padding: EdgeInsets.only(bottom: MediaQuery.of(context).viewInsets.bottom + 40),
                       child: FractionallySizedBox(
-                        heightFactor: 0.7,
+                        heightFactor: 0.8,
                         child: BlocProvider<PostBloc>.value(
                           value: postBloc,
                           child: CreateCommentModal(commentView: widget.commentViewTree),